# backend/services/hydration/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/entries": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all hydration entries for the user / Получить все записи пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hydration"
                ],
                "summary": "Get all hydration entries / Получить все записи",
                "responses": {
                    "200": {
                        "description": "List of hydration entries",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/hydration.HydrationEntry"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/hydration.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/hydration.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new hydration entry for the user / Добавить новую запись о приёме воды",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hydration"
                ],
                "summary": "Add hydration entry / Добавить запись о приёме воды",
                "parameters": [
                    {
                        "description": "Entry data / Данные записи",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hydration.CreateEntryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Entry created successfully",
                        "schema": {
                            "$ref": "#/definitions/hydration.HydrationEntry"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input",
                        "schema": {
                            "$ref": "#/definitions/hydration.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/hydration.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/hydration.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/goal": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update daily hydration goal for the user / Обновить дневную цель пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hydration"
                ],
                "summary": "Update daily goal / Обновить дневную цель",
                "parameters": [
                    {
                        "description": "Goal data / Новая цель",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hydration.UpdateGoalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Goal updated successfully",
                        "schema": {
                            "$ref": "#/definitions/hydration.UpdateGoalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input",
                        "schema": {
                            "$ref": "#/definitions/hydration.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/hydration.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/hydration.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get hydration statistics for the user / Получить статистику пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hydration"
                ],
                "summary": "Get hydration stats / Получить статистику",
                "responses": {
                    "200": {
                        "description": "Hydration statistics",
                        "schema": {
                            "$ref": "#/definitions/hydration.HydrationStats"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/hydration.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/hydration.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "hydration.CreateEntryRequest": {
            "type": "object",
            "required": [
                "amount",
                "type"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 250
                },
                "type": {
                    "type": "string",
                    "example": "water"
                }
            }
        },
        "hydration.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid input data"
                }
            }
        },
        "hydration.HydrationEntry": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 250
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "type": {
                    "type": "string",
                    "example": "water"
                },
                "user_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "hydration.HydrationStats": {
            "type": "object",
            "properties": {
                "goal": {
                    "type": "integer",
                    "example": 2000
                },
                "goal_percentage": {
                    "type": "integer",
                    "example": 75
                },
                "total_month": {
                    "type": "integer",
                    "example": 45000
                },
                "total_today": {
                    "type": "integer",
                    "example": 1500
                },
                "total_week": {
                    "type": "integer",
                    "example": 10500
                }
            }
        },
        "hydration.UpdateGoalRequest": {
            "type": "object",
            "required": [
                "goal"
            ],
            "properties": {
                "goal": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 2000
                }
            }
        },
        "hydration.UpdateGoalResponse": {
            "type": "object",
            "properties": {
                "goal": {
                    "type": "integer",
                    "example": 2000
                },
                "message": {
                    "type": "string",
                    "example": "Goal updated successfully"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}